+++ example/index.js
@@ -22,10 +22,6 @@ export const Main = () => {
     configPushNotification({
       notificationIcon: 'assets_images_calling_face',
       prefix: 'Cuộc gọi tới từ: ',
-      declineTitle: 'Từ chối',
-      acceptTitle: 'Chấp nhận',
-      acceptBackgroundColor: '#FF3700B3',
-      declineBackgroundColor: '#FF000000',
       incomingBackgroundColor: '#FFFFFFFF',
       incomingAcceptButtonImage: 'assets_images_join_call',
       incomingDeclineButtonImage: 'assets_images_hangup',